+++ tests/commands/test_options.py
@@ -92,3 +92,11 @@ def test_add_permission_option(parser: argparse.ArgumentParser):
     assert vars(parser.parse_args([])) == dict(permission=False)
     assert vars(parser.parse_args(["-p"])) == dict(permission=True)
     assert vars(parser.parse_args(["--permission"])) == dict(permission=True)
+
+
+def test_add_jobs_option(parser: argparse.ArgumentParser):
+    options.add_jobs_option(parser)
+
+    assert vars(parser.parse_args([])) == dict(jobs=1)
+    assert vars(parser.parse_args(["-j", "4"])) == dict(jobs=4)
+    assert vars(parser.parse_args(["--jobs", "0"])) == dict(jobs=0)